Added Animation
Added animation with Jim Sock.  The switch from a 3D object to a 2D sprite produces multiple problems such as the fact that looking at the object from the side makes it appear like you are looking in the direction you are traveling.  In conjonction with flipping the sprite (part of the animations) this really confuses the player as to what direction Jim Sock is actually going.  The large size of Jim Sock and the flat collider are also issues.  Long story short, when I made the camera I thought that our character was going to be in 3D.